+++ juniper-from-schema/src/lib.rs
@@ -1431,7 +1431,7 @@
     unused_variables,
     unused_must_use
 )]
-#![doc(html_root_url = "https://docs.rs/juniper-from-schema/0.5.0")]
+#![doc(html_root_url = "https://docs.rs/juniper-from-schema/0.5.1")]
 
 use juniper::{DefaultScalarValue, LookAheadSelection};
 use std::marker::PhantomData;